Change relevant URLs from "campaigns" to "projects"
