+++ src/features/surveys/components/EmptyOverview.tsx
@@ -22,7 +22,7 @@ const EmptyOverview = ({ campId, orgId, surveyId }: EmptyOverviewProps) => {
         <Msg id={messageIds.overview.noQuestions.title} />
       </Typography>
       <Link
-        href={`/organize/${orgId}/campaigns/${campId}/surveys/${surveyId}/questions`}
+        href={`/organize/${orgId}/projects/${campId}/surveys/${surveyId}/questions`}
         sx={{ marginTop: 4 }}
         underline="none"
       >
