+++ src/pages/organize/[orgId]/projects/[campId]/activities/index.tsx
