+++ src/pages/organize/[orgId]/projects/calendar/index.tsx
@@ -89,7 +89,7 @@ const AllCampaignsCalendarPage: PageWithLayout<
         <title>{messages.layout.calendar()}</title>
       </Head>
       <Calendar
-        baseHref={`/organize/${orgId}/campaigns/calendar`}
+        baseHref={`/organize/${orgId}/projects/calendar`}
         campaigns={campaigns}
         events={events}
         tasks={tasks}
